requirements.txt
fastapi~=0.115.6
SQLAlchemy~=2.0.36
uvicorn~=0.34.0
psycopg2-binary~=2.9.10
pydantic~=2.10.4
bcrypt~=4.2.1
PyJWT~=2.10.1
python-multipart~=0.0.20
websockets~=14.2
starlette~=0.41.3